fix: address CodeRabbit nitpick comments
- Remove deprecated markdownlint-cli2.jsonc file
- Fix grammar in test fixture ('This tests' → 'This fixture tests')
- Simplify Prettier version detection logic
- Improve test module mocking setup order
- Make test assertions more deterministic
- Extract duplicated normalizeLanguage method to utils/language.ts
- Fix markdown formatting (add blank lines around code blocks)
- Clean up remaining any type usage in tests
