--- .markdownlint-cli2.jsonc
@@ -1,73 +1,0 @@
-{
-  "extends": null,
-  "default": true,
-  "MD003": {
-    "style": "atx"
-  },
-  "MD004": {
-    "style": "dash"
-  },
-  "MD007": {
-    "indent": 2
-  },
-  "MD013": false,
-  "MD024": {
-    "siblings_only": true
-  },
-  "MD029": {
-    "style": "ordered"
-  },
-  "MD033": false,
-  "MD041": false,
-  "MD043": false,
-  "MD046": {
-    "style": "fenced"
-  },
-  "ignores": [
-    "node_modules/**",
-    ".git/**",
-    "dist/**",
-    "build/**",
-    "coverage/**",
-    "CHANGELOG.md",
-    "*.min.md"
-  ],
-  "terminology": [
-    {
-      "incorrect": "NPM",
-      "correct": "npm"
-    },
-    {
-      "incorrect": "Javascript",
-      "correct": "JavaScript"
-    },
-    {
-      "incorrect": "Typescript",
-      "correct": "TypeScript"
-    },
-    {
-      "incorrect": "VSCode",
-      "correct": "VS Code"
-    },
-    {
-      "incorrect": "MacOS",
-      "correct": "macOS"
-    },
-    {
-      "incorrect": "Github",
-      "correct": "GitHub"
-    },
-    {
-      "incorrect": "gitlab",
-      "correct": "GitLab"
-    },
-    {
-      "incorrect": "nodejs",
-      "correct": "Node.js"
-    },
-    {
-      "incorrect": "react native",
-      "correct": "React Native"
-    }
-  ]
-}
