+++ packages/rightdown/src/__tests__/fixtures/markdown/nested-blocks.md
@@ -1,6 +1,6 @@
 # Nested Code Blocks Test
 
-This tests nested code blocks and edge cases.
+This fixture tests nested code blocks and edge cases.
 
 ## Markdown with Code Blocks
