+++ packages/rightdown/src/core/orchestrator.ts
@@ -4,6 +4,7 @@ import { RIGHTDOWN_ERROR_CODES } from './errors.js';
 import { type RightdownConfig } from './types.js';
 import { type IFormatter } from '../formatters/base.js';
 import { AstProcessor } from '../processors/ast.js';
+import { normalizeLanguage } from '../utils/language.js';
 
 export interface OrchestratorOptions {
   config: RightdownConfig;
@@ -188,15 +189,6 @@ export class Orchestrator {
    * Normalize language identifier
    */
   private normalizeLanguage(language: string): string {
-    const aliases: Record<string, string> = {
-      js: 'javascript',
-      ts: 'typescript',
-      jsx: 'javascript',
-      tsx: 'typescript',
-      yml: 'yaml',
-      md: 'markdown',
-    };
-
-    return aliases[language.toLowerCase()] || language.toLowerCase();
+    return normalizeLanguage(language);
   }
 }
